Refactoring of the Personalised Dashboard architeture
