+++ src/client/components/InvestmentProjectSummary/state.js
@@ -1,4 +1,4 @@
+export const ID = 'investmentProjectSummary'
 export const TASK_GET_INVESTMENT_SUMMARY_DATA_RANGES =
   'TASK_GET_INVESTMENT_SUMMARY_DATA_RANGES'
-export const ID = 'investmentSummaryDataRanges'
 export const state2props = (state) => state[ID]
