+++ src/client/components/InvestmentReminders/state.js
@@ -1,6 +1,4 @@
-export const ID = 'personalisedDashboard'
-
+export const ID = 'investmentReminders'
 export const TASK_GET_OUTSTANDING_PROPOSITIONS =
   'TASK_GET_OUTSTANDING_PROPOSITIONS'
-
 export const state2props = (state) => state[ID]
