+++ src/client/index.jsx
@@ -102,8 +102,8 @@ import { TASK_GET_MY_INVESTMENTS_LIST } from './components/MyInvestmentProjects/
 import { fetchInvestmentSummaryDataRanges } from './components/InvestmentProjectSummary/tasks'
 import { TASK_GET_INVESTMENT_SUMMARY_DATA_RANGES } from './components/InvestmentProjectSummary/state'
 
-import { fetchOutstandingPropositions } from './components/PersonalisedDashboard/tasks'
-import { TASK_GET_OUTSTANDING_PROPOSITIONS } from './components/PersonalisedDashboard/state'
+import { fetchOutstandingPropositions } from './components/InvestmentReminders/tasks'
+import { TASK_GET_OUTSTANDING_PROPOSITIONS } from './components/InvestmentReminders/state'
 
 import Footer from '../client/components/Footer'
